// src/lib/api/types/adminListPostReportsParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export type AdminListPostReportsParams = {
/**
 * Filter by status: pending, under_review, resolved, dismissed
 */
status?: string;
/**
 * Search in post content
 */
search?: string;
/**
 * Page number
 */
page?: number;
/**
 * Page size
 */
pageSize?: number;
};