// src/lib/api/types/blockUserResponseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface BlockUserResponseBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** Success message */
  message: string;
}